+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # CHANGELOG MODULE CLICHAUMEIL FOR [DOLIBARR ERP CRM](https://www.dolibarr.org)
 
+## Unreleased
+- NEW : External supplier proposal list/card show supplier label (ref_fourn) and hydration updated *19/01/2026* -1.11.0
+- NEW : Add script to add ref_fourn column to supplier proposals *19/01/2026* - 1.11.0
+
 ## 1.10
 - FIX: Load additional language file and adjust supplier proposal status handling *15/01/2026* - 1.10.0
 - NEW: Add user contact assignment to supplier proposals in triggers *15/01/2026* - 1.10.0